(SIMP-828) Updated acceptance test
Realized, while reviewing the new docs, that the existing acceptance
test used the same password everywhere and that that was not a good test
of the system since a flaw could have passed through.

Updated all of the passwords to moderately random, but specific,
garbage.

SIMP-828 #comment Updated acceptance tests