(SIMP-2003) Fixed the acceptance tests
The acceptance tests were failing both because the network interfaces
were not being re-enabled and because the SELinux contexts were not
being set properly in the YUM test.

The MCO test is still failing due to our MCO module not handling the AIO
package properly but that's not a problem with running the tests
inherently and will need additional work.

SIMP-2003 #comment Fixed acceptance tests