(SIMP-2742) Clarify sysctl's ipv6 local variables (#87)
Prior to this `simp::sysctl` used a local variable, `$_ipv6`, which held
the value to `net.ipv6.conf.all.disable_ipv6`.  That was confusing,
because the value is whether or not to disable IPv6, the inverse of
expected meaning.  This commit changes the name of the local variable to
`$_disable_ipv6`.

SIMP-2742 #close