(SIMP-2839) Add test for installing via r10k
Add basic acceptance test for installing SIMP from r10k and a
Puppetfile. It should set all required settings from simp_options and
other modules.

First, it uses the copied modules from beaker to set up a puppetserver
with SIMP's pupmod module. It then copies the provided Puppetfile to the
master, installs the r10k gem, and runs r10k puppetfile install. It
drops in a barebones site.pp with a few node definitions and then runs
puppet on the master and the agent.

SIMP-2839 #close