(SIMP-2911) Reimplement base class inclusions
Prior to this commit classes were included via special "scenario"
classes, and mixed with the classes from the top-level Hiera key. That
implementaiton was problematic for a few reasons:

  1. It prevented fine-grained exclusions of the scenario classes.
  2. It broke with general best practices of not using top-scope Hiera
     keys in published modules.
  3. It was rather large and overly complex.

This commit seeks to fix these problems by moving the scenarios into the
module data layer, and morphing the top-scop Hiera key into a class
parameter, `$simp::classes`.  Fine-grained control over the scenario
data is controlled via a knockout prefix, similar to the one built in to
Hiera, but implemented as a new function, `simp::knockout()`.

SIMP-2911 #close