SIMP-3512 changes to pupmod-simp-simp for puppet supported tags
Specific items addressed in this PR:
* several syntax changes, such as multi-lining code blocks
* using correct keywords removed use of the 'inherits' keyword, which should only be used with the old params pattern
* remove variables from private classes, instead refer to the toplevel class
* comments welcome, still testing.

In the future, we need to address the use of 'include' instead of 'require' in many classes of this module. Currently many 'include' statements cannot be replaced by 'require' without creating cyclical dependencies, which may speak to a larger problem with defined types.

SIMP-3512 #close