(SIMP-4482) Set ownership and permissions of puppet/puppetdb.conf
Set the ownership and permissions of puppet/puppetdb.conf in
simp::puppetdb, instead of allowing them to be set to those of
the process running puppet, if the file needs to be created.
This is part of the fix to the failure of SIMP to bootstrap on a
system on which root's umask has already been restricted to 077.

SIMP-4482 #comment pupmod-simp-simp fix for puppet/puppetdb.conf