(SIMP-4663) Pin default tmp_service to system state
When running in the wild, it was discovered that it is unlikely that
switching over to the tmp.mounts service will actually be able to
succeed.

The module was changed to set simp::mountpoints::tmp::tmp_service to
`false` by default and `true` if the existing `/tmp` mountpoint has a
filesystem type of `tmpfs`.

SIMP-4663 #close